+++ crates/common/rpc_clients/bss/src/client.rs
@@ -41,7 +41,7 @@ impl RpcClient {
         request_id: u32,
         frame: rpc_codec_common::MessageFrame<bss_codec::MessageHeader, bytes::Bytes>,
         timeout: Option<std::time::Duration>,
-        trace_id: Option<u64>,
+        trace_id: Option<u128>,
         operation: Option<crate::stats::OperationType>,
     ) -> Result<rpc_codec_common::MessageFrame<bss_codec::MessageHeader>, rpc_client_common::RpcError>
     {
@@ -68,7 +68,7 @@ impl RpcClient {
         request_id: u32,
         frame: rpc_codec_common::MessageFrame<bss_codec::MessageHeader, Vec<bytes::Bytes>>,
         timeout: Option<std::time::Duration>,
-        trace_id: Option<u64>,
+        trace_id: Option<u128>,
         operation: Option<crate::stats::OperationType>,
     ) -> Result<rpc_codec_common::MessageFrame<bss_codec::MessageHeader>, rpc_client_common::RpcError>
     {
