+++ crates/common/rpc_clients/client_common/src/lib.rs
@@ -131,7 +131,7 @@ where
         request_id: u32,
         frame: MessageFrame<Header, Bytes>,
         timeout: Option<Duration>,
-        trace_id: Option<u64>,
+        trace_id: Option<u128>,
     ) -> Result<MessageFrame<Header>, RpcError> {
         self.ensure_connected().await?;
         let client = {
@@ -148,7 +148,7 @@ where
         request_id: u32,
         frame: MessageFrame<Header, Vec<bytes::Bytes>>,
         timeout: Option<Duration>,
-        trace_id: Option<u64>,
+        trace_id: Option<u128>,
     ) -> Result<MessageFrame<Header>, RpcError> {
         self.ensure_connected().await?;
         let client = {
@@ -262,7 +262,7 @@ macro_rules! rss_rpc_retry {
     };
 }
 
-pub fn encode_protobuf<M: PbMessage>(msg: M, _trace_id: Option<u64>) -> Result<Bytes, RpcError> {
+pub fn encode_protobuf<M: PbMessage>(msg: M, _trace_id: Option<u128>) -> Result<Bytes, RpcError> {
     let mut msg_bytes = BytesMut::with_capacity(1024);
     msg.encode(&mut msg_bytes)
         .map_err(|e| RpcError::EncodeError(e.to_string()))?;
