+++ crates/common/rpc_clients/nss/src/client.rs
@@ -41,7 +41,7 @@ impl RpcClient {
         request_id: u32,
         frame: rpc_codec_common::MessageFrame<nss_codec::MessageHeader, bytes::Bytes>,
         timeout: Option<std::time::Duration>,
-        trace_id: Option<u64>,
+        trace_id: Option<u128>,
         operation: crate::stats::NssOperation,
     ) -> Result<rpc_codec_common::MessageFrame<nss_codec::MessageHeader>, rpc_client_common::RpcError>
     {
